Move controller section to a blueprint
