+++ app/__init__.py
@@ -28,4 +28,7 @@ app.register_blueprint(customer_bp)
 from app.order import bp as order_bp
 app.register_blueprint(order_bp)
 
+from app.controller import bp as controller_bp
+app.register_blueprint(controller_bp)
+
 from app import routes, models
