+++ app/controller/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('controller', __name__)
+
+from app.controller import routes, forms
