+++ app/templates/controller/detail_controller.html
