+++ app/templates/customer/detail_customer.html
@@ -23,8 +23,6 @@
         </thead>
     </table>
     <div class="text-center">
-        <a class="btn btn-default" href="#">New Controller</a>
-        <a class='btn btn-default' href="{{ url_for('create_order', customer_id=customer.id) }}">New Order</a>
-        <a class='btn btn-default' href="#">New Ride</a>
+        <a class='btn btn-default' href="{{ url_for('order.create_order', customer_id=customer.id) }}">New Order</a>
     </div>
 {% endblock %}
