+++ app/templates/customer/list_customers.html
@@ -23,8 +23,8 @@
                     <a href="{{ url_for('customer.detail_customer', number=customer.id) }}">{{ customer.name.title() }}</a>
                 </td>
                 <td><a class="btn btn-default"
-                       href="{{ url_for('customer_orders', customer_id=customer.id) }}">Orders</a></td>
-                <td><a class="btn btn-primary" href="{{ url_for('create_order', customer_id=customer.id) }}">New
+                       href="{{ url_for('order.customer_orders', customer_id=customer.id) }}">Orders</a></td>
+                <td><a class="btn btn-primary" href="{{ url_for('order.create_order', customer_id=customer.id) }}">New
                     Order</a></td>
                 <td><a class="btn btn-info" href="{{ url_for('customer.edit_customer', number=customer.id) }}">Edit Customer</a>
                 </td>
